+++ loaders.c
@@ -35,6 +35,9 @@ const char loaders_rcs[] = "$Id$";
  *
  * Revisions   :
  *    $Log$
+ *    Revision 1.39  2002/03/07 03:46:17  oes
+ *    Fixed compiler warnings
+ *
  *    Revision 1.38  2002/03/06 22:54:35  jongfoster
  *    Automated function-comment nitpicking.
  *
@@ -493,7 +496,7 @@ int check_file_changed(const struct file_list * current,
 jb_err simple_read_line(FILE *fp, char **dest, int *newline)
 {
    int len = 0;
-   int buflen = BUFFER_SIZE;
+   size_t buflen = BUFFER_SIZE;
    char * buf;
    char * p;
    int ch;
@@ -893,7 +896,7 @@ jb_err edit_read_line(FILE *fp,
  *                Otherwise, returns buf.
  *
  *********************************************************************/
-char *read_config_line(char *buf, int buflen, FILE *fp, unsigned long *linenum)
+char *read_config_line(char *buf, size_t buflen, FILE *fp, unsigned long *linenum)
 {
    jb_err err;
    char *buf2 = NULL;
@@ -909,7 +912,7 @@ char *read_config_line(char *buf, int buflen, FILE *fp, unsigned long *linenum)
    else
    {
       assert(buf2);
-      assert(strlen(buf2) + 1U < (unsigned)buflen);
+      assert(strlen(buf2) + 1U < buflen);
       strncpy(buf, buf2, buflen - 1);
       free(buf2);
       buf[buflen - 1] = '\0';
