+++ urlmatch.c
@@ -33,6 +33,9 @@ const char urlmatch_rcs[] = "$Id$";
  *
  * Revisions   :
  *    $Log$
+ *    Revision 1.4  2002/03/07 03:46:17  oes
+ *    Fixed compiler warnings
+ *
  *    Revision 1.3  2002/03/03 14:51:11  oes
  *    Fixed CLF logging: Added ocmd member for client's request to struct http_request
  *
@@ -299,7 +302,7 @@ jb_err parse_http_url(const char * url,
     */
    {
       char *vec[BUFFER_SIZE];
-      int size;
+      size_t size;
       char *p;
 
       http->dbuffer = strdup(http->host);
@@ -675,7 +678,7 @@ jb_err create_url_spec(struct url_spec * url, const char * buf)
    if (buf[0] != '\0')
    {
       char *v[150];
-      int size;
+      size_t size;
 
       /* Parse domain part */
       if (buf[strlen(buf) - 1] == '.')
